move decipher to pkg
